Revert change in ImageCache

+++ Squeezer/src/main/java/uk/org/ngo/squeezer/util/ImageCache.java
@@ -408,7 +408,7 @@ public class ImageCache {
 
         synchronized (mDiskCacheLock) {
             mDiskCacheStarting = true;
-            if (mDiskLruCache != null) {
+            if (mDiskLruCache != null && !mDiskLruCache.isClosed()) {
                 try {
                     mDiskLruCache.delete();
                     if (BuildConfig.DEBUG) {
